feat: 인메모리 Repository 이동 및 ReservationRepository 인터페이스화
